refactor: simplify pricing cards component by integrating PricingCard and enhancing subscription handling
